+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 - Added `DynamicRangeValidator` type detection to `AttributeTypecastBehavior`
 - Enhanced default mailer DSN settings (Issue #32)
+- Enhanced `LogGridView` by HTML encoding the message column (Issue #31)
 - Fixed `TrailModelCollection` to allow primary keys with "-" characters 
 
 ## 2.3.17 (Oct 2, 2024)